# src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  title: Exhibition Management System API
  description: API for managing exhibitions, modules, stalls, and user preferences.
  version: 1.0.0
servers:
  - url: /api/
paths:
...
  /admin/exhibitions:
    post:
      summary: Create a new exhibition
      operationId: createExhibition
      tags:
        - Admin
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewExhibition'
      responses:
        '201':
          description: Exhibition created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Exhibition'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden - not an admin
    get:
      summary: Retrieve all exhibitions
      operationId: getExhibitions
      tags:
        - Admin
      security:
        - BearerAuth: []
      responses:
        '200':
          description: A list of exhibitions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExhibitionList'
        '401':
          description: Unauthorized
  /admin/exhibitions/{exhibitionId}:
    get:
      summary: Retrieve a single exhibition by ID
      operationId: getExhibitionById
      tags:
        - Admin
      security:
        - BearerAuth: []
      parameters:
        - name: exhibitionId
          in: path
          required: true
          schema:
            type: string
          description: The ID of the exhibition to retrieve
      responses:
        '200':
          description: Exhibition details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Exhibition'
        '401':
          description: Unauthorized
        '404':
          description: Exhibition not found
    put:
      summary: Update an existing exhibition
      operationId: updateExhibition
      tags:
        - Admin
      security:
        - BearerAuth: []
      parameters:
        - name: exhibitionId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateExhibition'
      responses:
        '200':
          description: Exhibition updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Exhibition'
        '401':
          description: Unauthorized
        '404':
          description: Exhibition not found
    delete:
      summary: Delete an exhibition
      operationId: deleteExhibition
      tags:
        - Admin
      security:
        - BearerAuth: []
      parameters:
        - name: exhibitionId
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Exhibition deleted successfully
        '401':
          description: Unauthorized
        '404':
          description: Exhibition not found
...
  /exhibitions:
    get:
      summary: Retrieve a list of published exhibitions
      operationId: getPublishedExhibitions
      tags:
        - User
      security:
        - BearerAuth: []
      parameters:
        - name: location
          in: query
          required: false
          schema:
            type: string
          description: Filter exhibitions by location
      responses:
        '200':
          description: A list of published exhibitions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExhibitionList'
        '401':
          description: Unauthorized
  /exhibitions/{exhibitionId}:
    get:
      summary: Retrieve a single exhibition and its modules
      operationId: getPublishedExhibitionById
      tags:
        - User
      security:
        - BearerAuth: []
      parameters:
        - name: exhibitionId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Exhibition details with modules
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Exhibition'
        '401':
          description: Unauthorized
        '404':
          description: Exhibition not found
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ExhibitionList:
      type: array
      items:
        $ref: '#/components/schemas/Exhibition'
    Exhibition:
      type: object
      properties:
        exhibitionId:
          type: string
        name:
          type: string
        description:
          type: string
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        location:
          type: string
        status:
          type: string
          enum: [Draft, Published]
        modules:
          type: array
          items:
            $ref: '#/components/schemas/Module'
    NewExhibition:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        location:
          type: string
        status:
          type: string
          enum: [Draft, Published]
    UpdateExhibition:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        location:
          type: string
        status:
          type: string
          enum: [Draft, Published]
...
    Module:
      type: object
      properties:
        moduleId:
          type: string
        name:
          type: string
        description:
          type: string
        assignedTeamId:
          type: string
        stalls:
          type: array
          items:
            $ref: '#/components/schemas/Stall'
...
    Stall:
      type: object
      properties:
        stallId:
          type: string
        name:
          type: string
        description:
          type: string
        stallNumber:
          type: string
        layout:
          type: string
        posterContent:
          type: array
          items:
            $ref: '#/components/schemas/PosterContent'
...
    PosterContent:
      type: object
      properties:
        contentId:
          type: string
        languageCode:
          type: string
        posterMediaUrl:
          type: string
        contentText:
          type: string
...